Converts config file location tests to tap API in separate file.
