+++ test/unit/config/config.test.js
@@ -10,12 +10,10 @@ const tap = require('tap')
 // Below allows use of mocha DSL with tap runner.
 tap.mochaGlobals()
 
-const path = require('path')
 const chai = require('chai')
 const should = chai.should()
 const sinon = require('sinon')
 const expect = chai.expect
-const fs = require('fs')
 const Config = require('../../../lib/config')
 
 tap.test('should handle a directly passed minimal configuration', (t) => {
@@ -29,184 +27,6 @@ tap.test('should handle a directly passed minimal configuration', (t) => {
 })
 
 describe('the agent configuration', function() {
-  describe('when overriding the config file location via NR_HOME', function() {
-    var origHome = null
-    var startDir = null
-    var DESTDIR = path.join(__dirname, 'xXxNRHOMETESTxXx')
-    var NOPLACEDIR = path.join(__dirname, 'NOHEREHERECHAMP')
-    var CONFIGPATH = path.join(DESTDIR, 'newrelic.js')
-
-
-    beforeEach(function(done) {
-      if (process.env.NEW_RELIC_HOME) {
-        origHome = process.env.NEW_RELIC_HOME
-      }
-
-      startDir = process.cwd()
-
-      fs.mkdir(DESTDIR, function(error) {
-        if (error) return done(error)
-
-        fs.mkdir(NOPLACEDIR, function(error) {
-          if (error) return done(error)
-
-          process.chdir(NOPLACEDIR)
-          process.env.NEW_RELIC_HOME = DESTDIR
-
-          var sampleConfig = fs.createReadStream(
-            path.join(__dirname, '../../../lib/config/default.js')
-          )
-          var sandboxedConfig = fs.createWriteStream(CONFIGPATH)
-          sampleConfig.pipe(sandboxedConfig)
-
-          sandboxedConfig.on('close', function() { return done() })
-        })
-      })
-    })
-
-    afterEach(function(done) {
-      if (origHome) {
-        process.env.NEW_RELIC_HOME = origHome
-      } else {
-        delete process.env.NEW_RELIC_HOME
-      }
-      origHome = null
-
-      fs.unlink(CONFIGPATH, function(error) {
-        if (error) return done(error)
-
-        fs.rmdir(DESTDIR, function(error) {
-          if (error) return done(error)
-
-          process.chdir(startDir)
-
-          fs.rmdir(NOPLACEDIR, done)
-        })
-      })
-    })
-
-    it('should load the configuration', function() {
-      expect(function() { Config.initialize() }).not.throws()
-    })
-
-    it('should export the home directory on the resulting object', function() {
-      var configuration = Config.initialize()
-      expect(configuration.newrelic_home).equal(DESTDIR)
-    })
-
-    it('should ignore the configuration file completely when so directed', function() {
-      process.env.NEW_RELIC_NO_CONFIG_FILE = 'true'
-      process.env.NEW_RELIC_HOME = '/xxxnoexist/nofile'
-
-      var configuration
-      expect(function envTest() {
-        configuration = Config.initialize()
-      }).not.throws()
-
-      should.not.exist(configuration.newrelic_home)
-      expect(configuration.error_collector &&
-             configuration.error_collector.enabled).equal(true)
-
-      delete process.env.NEW_RELIC_NO_CONFIG_FILE
-      delete process.env.NEW_RELIC_HOME
-    })
-  })
-
-  describe('Selecting config file path', () => {
-    let origHome
-    let originalWorkingDirectory
-    const DESTDIR = path.join(__dirname, 'test_NEW_RELIC_CONFIG_FILENAME')
-    const NOPLACEDIR = path.join(__dirname, 'test_NEW_RELIC_CONFIG_FILENAME_dummy')
-    const MAIN_MODULE_DIR = path.join(__dirname, 'test_NEW_RELIC_CONFIG_FILENAME_MAIN_MODULE')
-    let CONFIG_PATH
-    let processMainModuleStub
-
-    beforeEach(() => {
-      if (process.env.NEW_RELIC_HOME) {
-        origHome = process.env.NEW_RELIC_HOME
-      }
-
-      process.env.NEW_RELIC_HOME = DESTDIR
-
-      originalWorkingDirectory = process.cwd()
-
-      processMainModuleStub = sinon.stub(process, 'mainModule').value({
-        filename: `${MAIN_MODULE_DIR}/index.js`
-      })
-
-      fs.mkdirSync(DESTDIR)
-      fs.mkdirSync(NOPLACEDIR)
-      fs.mkdirSync(MAIN_MODULE_DIR)
-
-      process.chdir(NOPLACEDIR)
-    })
-
-    afterEach(() => {
-      if (origHome) {
-        process.env.NEW_RELIC_HOME = origHome
-      } else {
-        delete process.env.NEW_RELIC_HOME
-      }
-      origHome = null
-
-      if (CONFIG_PATH) {
-        fs.unlinkSync(CONFIG_PATH)
-        CONFIG_PATH = undefined
-      }
-
-      processMainModuleStub.resetBehavior()
-
-      fs.rmdirSync(DESTDIR)
-      fs.rmdirSync(NOPLACEDIR)
-      fs.rmdirSync(MAIN_MODULE_DIR)
-
-      process.chdir(originalWorkingDirectory)
-    })
-
-    const createSampleConfig = (dir, filename) => {
-      CONFIG_PATH = path.join(dir, filename)
-
-      const config = {
-        app_name: filename
-      }
-
-      fs.writeFileSync(CONFIG_PATH, `exports.config = ${JSON.stringify(config)}`)
-    }
-
-    it('should load the default newrelic.js config file', function() {
-      const filename = "newrelic.js"
-      createSampleConfig(DESTDIR, filename)
-
-      const configuration = Config.initialize()
-      expect(configuration.app_name).equal(filename)
-    })
-
-    it('should load the default newrelic.cjs config file', function() {
-      const filename = "newrelic.cjs"
-      createSampleConfig(DESTDIR, filename)
-
-      const configuration = Config.initialize()
-      expect(configuration.app_name).equal(filename)
-    })
-
-    it('should load config when overriding the default with NEW_RELIC_CONFIG_FILENAME', function() {
-      const filename =  'some-file-name.js'
-      process.env.NEW_RELIC_CONFIG_FILENAME = filename
-      createSampleConfig(DESTDIR, filename)
-
-      const configuration = Config.initialize()
-      expect(configuration.app_name).equal(filename)
-    })
-
-    it('should load config from the main module\'s filepath', function() {
-      const filename = 'newrelic.js'
-      createSampleConfig(MAIN_MODULE_DIR, filename)
-
-      const configuration = Config.initialize()
-      expect(configuration.app_name).equal(filename)
-    })
-  })
-
   describe('when loading invalid configuration file', function() {
     let realpathSyncStub
     const fsUnwrapped = require('../../../lib/util/unwrapped-core').fs
